bash: fix sending error logs and zips

+++ build.sh
@@ -11,6 +11,7 @@ MAINPATH=/home/timisong # измените, если необходимо
 
 # Каталог ядра
 KERNEL_DIR=$MAINPATH/kernel
+KERNEL_PATH=$KERNEL_DIR/kernel_xiaomi_sm8250
 
 HASH=$(git rev-parse HEAD)
 value=$(<../info.txt)
@@ -129,23 +130,22 @@ find $DTS -name 'dtbo.img' -exec cat {} + > $DTBOPATH
 cd "$MAGIC_TIME_DIR"
 7z a -mx9 MagicTime-$MODEL-$MAGIC_BUILD_DATE.zip * -x!*.zip
 
-curl -F document=@"./MagicTime-$MODEL-$MAGIC_BUILD_DATE.zip" -F caption="$HASH" "https://api.telegram.org/bot$TOKEN/sendDocument?chat_id=@magictimec"
-
-rm -rf MagicTime-$MODEL-$MAGIC_BUILD_DATE.zip
-
 # Завершение отсчета времени выполнения скрипта
 end_time=$(date +%s)
 elapsed_time=$((end_time - start_time))
 
     # Проверка успешности сборки
-    if [ $? -eq 0 ]; then
-        echo "\e[32mОбщее время выполнения: $elapsed_time секунд\e[0m"
-        curl -s -X POST "https://api.telegram.org/bot$TOKEN/sendMessage" -d chat_id="@magictimec" -d text="Компиляция завершилась успешно! Время выполенения: $elapsed_time секунд"
+    if [ $? -ne 0 ]; then
+        if [ $? -eq 0 ]; then
+            echo "\e[32mОбщее время выполнения: $elapsed_time секунд\e[0m"
+            curl -s -X POST "https://api.telegram.org/bot$TOKEN/sendMessage" -d chat_id="@magictimec" -d text="Компиляция завершилась успешно! Время выполенения: $elapsed_time секунд"
+            curl -F document=@"./MagicTime-$MODEL-$MAGIC_BUILD_DATE.zip" -F caption="$HASH" "https://api.telegram.org/bot$TOKEN/sendDocument?chat_id=@magictimec"
+            rm -rf MagicTime-$MODEL-$MAGIC_BUILD_DATE.zip
+        fi
     else
+        cd "$KERNEL_PATH"
         echo "\e[31mОшибка: Сборка завершилась с ошибкой\e[0m"
         curl -s -X POST "https://api.telegram.org/bot$TOKEN/sendMessage" -d chat_id="@magictimec" -d text="Ошибка в компиляции!"
         curl -F document=@"./error.log" "https://api.telegram.org/bot$TOKEN/sendDocument?chat_id=@magictimec"
     fi
-else
-    message="\e[31mОшибка: Не удалось настроить конфигурацию ядра\e[0m"
 fi
